Add support for manylinux_${GLIBCMAJOR}_${GLIBCMINOR} wheels
https://www.python.org/dev/peps/pep-0600/

This fixes e.g. using the latest tokenizers package (0.10.3) with
mach-nix.

TODO: Update WheelDependencyProvider to process all filename tags
individually instead of matching them all with a single regexp. If it
had done that already, this change would have been somewhat unneeded, as
all wheel files seem to be doubly tagged (at least for now):

* manylinux_2_5 is also tagged manylinux1
* manylinux_2_12 is also tagged manylinux2010
* manylinux_2_17 is also tagged manylinux2014

(But since the new format is first in the filename, the old regexps
don't match.)