Fix `nix (build|shell) mach-nix#gen`
`nix (build|shell) mach-nix#gen` is broken on recent nix versions because
of https://github.com/NixOS/nix/issues/6690.

This commit fixes it by removing `meta` from the list of handled packages.
As a result, nix can access the derivation's meta as expected, and work
as expected.

The side effect is that the `meta` package on pypi cannot be used, but
that's a cost we have to pay.